Reorder files - separate helper functions
